requirements.txt
fastapi~=0.111.1
uvicorn
firebase-admin
pydantic~=2.8.2

python-dotenv~=1.0.1
pip~=23.2.1
cryptography~=43.0.0
Jinja2~=3.1.4